% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_psodds.R
\name{add_multipliers}
\alias{add_multipliers}
\title{Generate a column of multipliers from the odds provided by the user in the format
specified in the package-wide manfile}
\usage{
add_multipliers(race_df)
}
\description{
Generate a column of multipliers from the odds provided by the user in the format
specified in the package-wide manfile
}

