% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{box}
\alias{box}
\title{Calculates the probability of a box bet being successful}
\usage{
box(race_df, picks)
}
\arguments{
\item{race_df, }{a data frame containing each horse's name and odds of winning 1st}

\item{picks}{character vector containing the better's picks for 1st, 2nd, and possibly third
picks in no particular order}
}
\value{
our calculation of probability
}
\description{
Calculates the probability of a box bet being successful
}
\examples{
box(Kentucky_derby, c('Nyquist','Exaggerator'))
box(race, my_picks)
box(race, c('Nyquist', 'Exagerator','Big Brown'))
}

