% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrameCreation.R
\name{create_df}
\alias{create_df}
\title{Creates the initial data frame that the user will add to later.}
\usage{
create_df()
}
\value{
data frame with horse names, odds for the race, and the race name
}
\description{
Creates the initial data frame that the user will add to later.
}
\examples{
create_df()


}

