% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fill_df.R
\name{fill_data}
\alias{fill_data}
\title{Fill the data frame inputted data, and assign the race and ID}
\usage{
fill_data(horse_names, odds, race_name)
}
\arguments{
\item{vec1}{is a vector of names}

\item{vec2}{is a numeric vector of odds}

\item{race}{is the name of the race}
}
\value{
the dataframe with the names of the horses and their odds within it
}
\description{
Fill the data frame inputted data, and assign the race and ID
}
\examples{
ellipse_perimeter(1, 1)
ellipse_perimeter(1)
}

