% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_odds.R
\name{plot_odds}
\alias{plot_odds}
\title{Plot horse's odds of winning first, second, and third, or showing}
\usage{
plot_odds(race_df, picks)
}
\arguments{
\item{race_df, }{a data frame that has been passed through generate_psodds, so it has columns
for each horse's odds of winning 1st, 2nd, and 3rd, or showing}

\item{picks, }{a character containing the horses whose odds of winning first, second, and third you want to see}
}
\value{
a ggplot for each horse
}
\description{
Plot horse's odds of winning first, second, and third, or showing
}
\examples{
exacta(Kentucky_derby, c('Nyquist','Exaggerator'))
exacta(race, my_picks)
}

