% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_psodds.R
\name{psodds_first}
\alias{psodds_first}
\title{Generate a column of the odds of each horse winning from the Multiplier column}
\usage{
psodds_first(race_df)
}
\description{
Generate a column of the odds of each horse winning from the Multiplier column
}

