% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_psodds.R
\name{psodds_second}
\alias{psodds_second}
\title{Generate a column of the odds of each horse coming in second. Same reasoning as
psodds_first.}
\usage{
psodds_second(race_df)
}
\description{
Generate a column of the odds of each horse coming in second. Same reasoning as
psodds_first.
}

