% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_psodds.R
\name{psodds_third}
\alias{psodds_third}
\title{Generate a column of the odds of each horse coming in third.}
\usage{
psodds_third(race_df)
}
\description{
Generate a column of the odds of each horse coming in third.
}

