% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win.R
\name{win}
\alias{win}
\title{Calculate the}
\usage{
win(race_df, pick, amount = 1)
}
\arguments{
\item{pick}{is the name of the horse the user wants to bet on}

\item{amount}{is the amount the user wants to bet.}
}
\value{
A vector of the payoff for that bet and the odds the user wins the bet
}
\description{
Calculate the
}
\examples{
ellipse_perimeter(1, 1)
ellipse_perimeter(1)
}

