% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exacta_bet.R
\name{exacta}
\alias{exacta}
\title{Calculates the probability of an exacta bet being successful}
\usage{
exacta(race_df, picks)
}
\arguments{
\item{race_df, }{a data frame containing each horse's name and odds of winning 1st}

\item{picks}{character vector containing the picks, by horse name, to come in 1st and 2nd
in that order.}
}
\value{
The probability of the exacta bet being successful.
}
\description{
Calculates the probability of an exacta bet being successful
}
\examples{
exacta(Kentucky_derby, c('Nyquist','Exaggerator'))
exacta(race, my_picks)
}

