% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_bet.R
\name{show_bet}
\alias{show_bet}
\title{Calculate the odds of a show bet}
\usage{
show_bet(race_df, pick)
}
\arguments{
\item{race_df, }{a data frame containing each horse's name and odds of winning 1st}

\item{Pick}{for the show bet}
}
\value{
our calculation of probability
}
\description{
Calculate the odds of a show bet
}
\examples{
show_bet(horse_info, "Nyquist")
}

